requirements.txt
aenum==3.1.15
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
appdirs==1.4.4
asgiref==3.8.1
attrs==23.2.0
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
ChatterBot==1.0.4
chatterbot-corpus==1.2.0
click==8.1.7
colorama==0.4.6
Deprecated==1.2.14
distlib==0.3.8
Django==5.0.6
filelock==3.15.1
Flask==3.0.3
flexcache==0.3
flexparser==0.3.1
frozenlist==1.4.1
future==1.0.0
greenlet==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
line-bot-sdk==3.11.0
MarkupSafe==2.1.5
mathparse==0.1.2
mongoengine==0.28.2
multidict==6.0.5
nltk==3.8.1
pbr==6.0.0
Pint==0.24
pipenv==2024.0.1
pipenv-to-requirements==0.9.0
platformdirs==4.2.2
pydantic==2.7.4
pydantic_core==2.18.4
pymongo==3.13.0
python-dateutil==2.7.5
pytz==2024.1
PyYAML==3.13
regex==2024.5.15
requests==2.31.0
setuptools==70.0.0
six==1.16.0
sqlparse==0.5.0
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
virtualenv==20.26.2
Werkzeug==3.0.3
wrapt==1.16.0
yarl==1.9.4
